backend/requirements.txt
# Web フレームワーク
fastapi
uvicorn[standard]

# ファイルアップロード／環境変数
python-multipart
python-dotenv

# XML パース
lxml

# PDF → 図表抽出
pdfplumber
tabula-py   # （要 JPype1: pip install JPype1）
pandas
Pillow      # pdfplumber で画像を扱う場合に必要

# HTTP クライアント
requests

# CERMINE から出力される XML を扱うために
# （tei2json.py 内で使っている lxml と組み合わせて使えます）

# 認証系・クラウド連携
openai
boto3
google-cloud-storage>=2.0.0
google-cloud-firestore>=2.0.0
google-api-python-client>=2.0.0
google-cloud-secret-manager>=2.0.0